+++ frontend/src/components/App/Participants/index.tsx
@@ -12,7 +12,6 @@ import useTranslation from 'next-translate/useTranslation';
 
 import { Participant } from '@/types/participant';
 import { pushEventDataLayer } from '@/lib/analytics';
-import { ping } from '@/user/auth';
 import { useJitsi } from '@/lib/useJitsi';
 import ParticipantCard from '@/components/App/Participants/ParticipantCard';
 
@@ -28,6 +27,7 @@ import ButtonCopyUrl from '@/components/Common/ButtonCopyUrl';
 import { useStooa } from '@/contexts/StooaManager';
 import { getApiParticipantList } from '@/repository/ApiParticipantRepository';
 import { useModals } from '@/contexts/ModalsContext';
+import { useUserAuth } from '@/user/auth/useUserAuth';
 
 const initialParticipant: Participant = {
   id: '',
@@ -53,6 +53,8 @@ const PING_TIMEOUT = 3500;
 
 const Participants: React.FC<Props> = ({ initialized, fid }) => {
   const { getParticipantList } = useJitsi();
+  const { ping } = useUserAuth();
+
   const { t, lang } = useTranslation('fishbowl');
   const pingInterval = useRef<number>();
   const getParticipantsInterval = useRef<number>();
