+++ frontend/src/lib/apollo-client.js
@@ -9,7 +9,7 @@
 
 import { ApolloClient, ApolloProvider, createHttpLink, InMemoryCache } from '@apollo/client';
 import { setContext } from '@apollo/client/link/context';
-import { getAuthToken } from '@/user/auth';
+import { getAuthToken } from '@/user/auth/useUserAuth';
 import LocaleCookie from '@/lib/LocaleCookie';
 
 const httpLink = createHttpLink({
